Fixed month bug? Added swirtch returns
The .tm_mon TIME structure variable seems to return a month number one
too low (Feb = 1). Maybe it's 0 indexed but that seems like a weird
design choice. Added "return" to switch cases so that the rest of the
Test App doesn't run if you only want to preform Unit tests.