Added exact command to load the UVCstill module.
Load the uvcstill module, by running the command "sudo insmod uvcstill.ko" in the Linux or Windows Cygwin terminal. THIS IS THE MOST IMPORTANT STEP. It needs to be done only once every time you are about to record for the first time or connect everything. IF you unplug a camera, you need to run this again before recording.